refactor: StringUtils class의 패키지 이동
