+++ src/test/java/util/StringUtilsTest.java
@@ -1,4 +1,4 @@
-package step1;
+package util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
